requirements.txt
numpy==2.2.6
pillow==12.0.0
protobuf
transformers
accelerate
ultralytics
langchain-chroma
langchain-huggingface
langchain-core
sentence-transformers
torch
torchvision
torchaudio
bitsandbytes
opencv-python-headless
datasets
scipy
tqdm
qwen_vl_utils